iris/sumi: registry startAlgOverride helper + collective fixups
Extract the param>env>default algorithm selection into
CollectiveEngine::startAlgOverride so the scalar-count engine ops share one
flat single-DAG override path (always wins, even when startCollective returns
nullptr for an async multi-rank collective).

Also fold two review fixes on the engine collectives:
- ring_allreduce.cc finalizeBuffers: widen nelems_*type_size_ to long before
  the multiply (matches initBuffers; int product could overflow large msgs).
- reduce.h toString(): "sumi allreduce" -> "sumi reduce" (copy-paste).

+++ src/sst/elements/iris/sumi/ring_allreduce.cc
@@ -26,7 +26,7 @@ namespace SST::Iris::sumi {
 void
 RingAllreduceActor::finalizeBuffers()
 {
-  long buffer_size = nelems_ * type_size_;
+  long buffer_size = static_cast<long>(nelems_) * type_size_; // widen before multiply, matches initBuffers
   my_api_->freeWorkspace(recv_buffer_, buffer_size);
   //send buffer aliases the result buffer
 }
